Updated glsl-optimizer.

+++ 3rdparty/glsl-optimizer/src/glsl/ast_function.cpp
@@ -1484,7 +1484,7 @@ ast_function_expression::hir(exec_list *instructions,
 	    var->constant_value = matrix->constant_expression_value();
 
 	    /* Replace the matrix with dereferences of its columns. */
-	    for (int i = 0; i < matrix->type->matrix_columns; i++) {
+	    for (int i = 0; i < (int)matrix->type->matrix_columns; i++) {
 	       matrix->insert_before(new (ctx) ir_dereference_array(var,
 		  new(ctx) ir_constant(i)));
 	    }
